# docs/swagger.yaml
definitions:
  models.Clip:
    properties:
      creator:
        $ref: '#/definitions/models.Creator'
      creator_uuid:
        type: string
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      game:
        $ref: '#/definitions/models.Game'
      game_uuid:
        type: string
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
      viewcount:
        type: integer
      vod:
        $ref: '#/definitions/models.Vod'
      vod_offset:
        type: integer
      vod_uuid:
        type: string
    type: object
  models.Creator:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
...
  models.Game:
    properties:
      box_art_url:
        type: string
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
  models.Vod:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      fps:
        type: number
      publish:
        type: boolean
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
paths:
  /clips/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a clip
        in: query
        name: uuid
        type: string
      - description: The title of a clip
        in: query
        name: title
        type: string
      - description: The duration of a clip
        in: query
        name: duration
        type: integer
      - description: The date of a clip
        in: query
        name: date
        type: string
      - description: The filename of a clip
        in: query
        name: filename
        type: string
      - description: The resolution of a clip
        in: query
        name: resolution
        type: string
      - description: The size of a clip
        in: query
        name: size
        type: integer
      - description: The viewcount of a clip
        in: query
        name: viewcount
        type: integer
      - description: The creator id of a clip
        in: query
        name: creator
        type: integer
      - description: The game id of a clip
        in: query
        name: game
        type: integer
      - description: The vod id of a clip
        in: query
        name: vod
        type: string
      - description: 'Set order direction divided by comma. Possible ordering values:
          ''date'', ''duration'', ''size''. Possible directions: ''asc'', ''desc''.
          Example: ''date,desc'''
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Clip'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all clips
      tags:
      - Clips
    post:
      consumes:
      - application/json
      parameters:
      - description: Clip obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Clip'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create clip
      tags:
      - Clips
  /clips/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete clip
      tags:
      - Clips
    get:
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Clip'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get clips by uuid
      tags:
      - Clips
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      - description: Clip obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Clip'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Patch clip
      tags:
      - Clips
  /creators/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a creator
        in: query
        name: uuid
        type: integer
      - description: The name of a creator
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Creator'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all creators
      tags:
      - Creators
    post:
      consumes:
      - application/json
      parameters:
      - description: Creator obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Creator'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create creator
      tags:
      - Creators
  /creators/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete creator
      tags:
      - Creators
    get:
      parameters:
      - description: Unique Identifyer
        in: path
        name: uuid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Creator'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get creator by uuid
      tags:
      - Creators
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: integer
      - description: Creator obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Creator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Patch creator
      tags:
      - Creators
...
swagger: "2.0"